Added utilities file
